requirements.txt
datasets==2.21.0
einops==0.8.0
evaluate==0.4.2
fire==0.6.0
flash_attn==2.6.1
huggingface_hub==0.23.4
matplotlib==3.9.2
numpy==1.26.4
parameterized==0.9.0
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.1
seaborn==0.13.2
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
wandb==0.17.7
